+++ src/main/kotlin/io/github/jamalam/extensions/user/UserUtilityExtension.kt
@@ -69,7 +69,8 @@ class UserUtilityExtension : Extension() {
 
         event<ThreadChannelCreateEvent> {
             action {
-                if (event.channel.guild.getConfig().moderationConfig.autoSaveThreads) {
+                if (event.channel.guild.getConfig().moderationConfig.autoSaveThreads &&
+                        event.channel.member == null) { // Only thread creation, not join
                     event.channel.save(true)
 
                     event.channel.guild.getLogChannel()?.createEmbed {